Added composite formatting to console outputs. Added comments.

+++ src/ActivityTracker/Activity.cs
@@ -1,17 +1,19 @@
 using System;
-using System.Collections.Generic;
 
 namespace ActivityTracker
-{
+{ 
+    //Activity is the object that users log
     public class Activity
     {
-
+        //When user logs activity, it is confirmed with the following method
         public void SummarizeEntry()
         {
             System.Console.WriteLine($"Your activity, {Name} has been logged.");
         }
+
+        //Activity fields
         public string Name {get; set;}
-        public int Duration {get; set;}
+        public double Duration {get; set;}
         public double PerceivedExertion {get; set;}
         public DateTime timeStamp {get; set;}
