+++ src/ActivityTracker/MainMenu.cs
@@ -10,13 +10,15 @@ namespace ActivityTracker
 
         ActivityLog activityLog = new ActivityLog();
 
+        //Load activity objects in activity.json for use
         public MainMenu()
         {
             activityLog.LoadActivities(path);
         }
+
+        //Create main menu and display options
         internal void DisplayMenu()
         {
-
             var quitMenu = false;
             while (!quitMenu)
             {
@@ -58,10 +60,11 @@ namespace ActivityTracker
                     }
                 }
             }
-            
+            //Save activities to activity.json
             activityLog.SaveActivities(path);
         }
         
+        //Called if user selects "1" from main menu. Creates new activity object, records user inputs and adds to ActivitiesList
         private void LogActivity()
         {
             Activity activity = new Activity();
@@ -69,7 +72,7 @@ namespace ActivityTracker
             var input = Console.ReadLine();
             activity.Name = input;
             System.Console.WriteLine("How long (minutes) did you perform the activity?");
-            activity.Duration = int.Parse(Console.ReadLine());
+            activity.Duration = double.Parse(Console.ReadLine());
             System.Console.WriteLine("What was your percieved exertion (1-10)?");
             activity.PerceivedExertion = double.Parse(Console.ReadLine()); 
             activity.timeStamp = DateTime.Now;
@@ -77,6 +80,7 @@ namespace ActivityTracker
             activity.SummarizeEntry();
         }
 
+        //Called if user selects "2" from main menu. User is presented more options. If user selects "1", activities list is summarized from the past week and displayed. If user selects "2", all activities from activities list are summarized.
         private void ListActivitySummary()
         {
             System.Console.WriteLine("Which activity summary would you like to see?");
@@ -108,6 +112,7 @@ namespace ActivityTracker
             
         }
 
+        //Called if user selects "3". Method clears activities from ActivitiesList. User has to confirm.
         private void ClearActivityLog()
         {
             System.Console.WriteLine("Are you sure you want to clear activity logs?");
